Enable health report on main master
TODO: when backup master becomes main master, open the health report.
ISSUE: Adding/Deleting Minion operations are racing with health_report. We need
a lock on minion related variables. Futhermore, we might need a faster health
report (multithread). Since the time for health_report is in linear relationship
with the amount of minion.